EVM: split rpc/mod.rs into separate files
